# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List active leads with optional server-side filtering
      description: Returns paginated active leads. Supports search, status filter, limit and offset.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /crm/leads:
    get:
      operationId: listCrmLeads
      tags: [crm]
      summary: List CRM leads within the caller scope
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [rep, manager]
            default: rep
        - name: status
          in: query
          schema:
            type: string
        - name: sales_rep
          in: query
          schema:
            type: string
        - name: funnel
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: CRM leads
          headers:
            X-Total-Count:
              description: Total number of CRM leads matching the requested filters and scope
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrmLead"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCrmLead
      tags: [crm]
      summary: Create a manually entered CRM lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLeadInput"
      responses:
        "201":
          description: Created CRM lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLead"
        "400":
          description: Invalid CRM lead data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /crm/leads/{id}:
    get:
      operationId: getCrmLead
      tags: [crm]
      summary: Get a CRM lead within the caller scope
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [rep, manager]
            default: rep
      responses:
        "200":
          description: CRM lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLead"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found in the caller scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCrmLead
      tags: [crm]
      summary: Update an editable CRM lead field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [rep, manager]
            default: rep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrmLeadUpdate"
      responses:
        "200":
          description: Updated CRM lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmLead"
        "400":
          description: Invalid CRM lead update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found in the caller scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Lead:
      type: object
      properties:
        id:
          type: string
        lead_number:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        phone_link:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        salesperson_id:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        answer_status:
          type: string
          nullable: true
        capture_attempt_status:
          type: string
          nullable: true
        lead_source:
          type: string
          nullable: true
        ad_name:
          type: string
          nullable: true
        referral_name:
          type: string
          nullable: true
        activity_field:
          type: string
          nullable: true
        followup_at:
          type: string
          nullable: true
        followup_note:
          type: string
          nullable: true
        reminder_at:
          type: string
          nullable: true
        reminder_note:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        rejection_reason_text:
          type: string
          nullable: true
        lead_created_at:
          type: string
          nullable: true
        closed_at:
          type: string
          nullable: true
        first_deal_amount:
          type: string
          nullable: true
        lead_count:
          type: integer
          nullable: true
        linked_customer_id:
          type: string
          nullable: true
        task_text:
          type: string
          nullable: true
        task_due_at:
          type: string
          nullable: true
        monday_board_id:
          type: string
          nullable: true
        monday_item_id:
          type: string
          nullable: true
        monday_group_id:
          type: string
          nullable: true
        created_at:
          type: string
          nullable: true
        updated_at:
          type: string
          nullable: true
        deleted_at:
          type: string
          nullable: true
      required:
        - id
        - lead_number
        - name
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        answer_status:
          type: string
          nullable: true
        capture_attempt_status:
          type: string
          nullable: true
        lead_source:
          type: string
          nullable: true
        ad_name:
          type: string
          nullable: true
        referral_name:
          type: string
          nullable: true
        activity_field:
          type: string
          nullable: true
        followup_at:
          type: string
          nullable: true
        followup_note:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        rejection_reason_text:
          type: string
          nullable: true
        first_deal_amount:
          type: string
          nullable: true
        task_text:
          type: string
          nullable: true
      required:
        - name
    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        answer_status:
          type: string
          nullable: true
        lead_source:
          type: string
          nullable: true
        ad_name:
          type: string
          nullable: true
        referral_name:
          type: string
          nullable: true
        activity_field:
          type: string
          nullable: true
        followup_at:
          type: string
          nullable: true
        followup_note:
          type: string
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        rejection_reason_text:
          type: string
          nullable: true
        first_deal_amount:
          type: string
          nullable: true
        task_text:
          type: string
          nullable: true
...
    CrmLead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone_e164:
          type: string
        phone_raw:
          type: [string, "null"]
        email:
          type: [string, "null"]
        sales_rep_id:
          type: [string, "null"]
        status_code:
          type: string
        is_active_customer:
          type: boolean
        answer_status:
          type: [string, "null"]
        capture_attempts:
          type: integer
        rejection_reason_code:
          type: [string, "null"]
        rejection_detail:
          type: [string, "null"]
        pending_reassignment:
          type: boolean
        legacy_lead_id:
          type: [string, "null"]
        linked_customer_id:
          type: [string, "null"]
        source:
          type: string
        source_ref:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: [string, "null"]
          format: date-time
      required:
        - id
        - name
        - phone_e164
        - status_code
        - is_active_customer
        - capture_attempts
        - pending_reassignment
        - source
        - created_at
        - updated_at
    CrmLeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: [string, "null"]
        answer_status:
          type: [string, "null"]
        rejection_reason_code:
          type: [string, "null"]
        rejection_detail:
          type: [string, "null"]
        source_ref:
          type: [string, "null"]
      required: [name, phone]
    CrmLeadUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: [string, "null"]
        answer_status:
          type: [string, "null"]
        rejection_reason_code:
          type: [string, "null"]
        rejection_detail:
          type: [string, "null"]
        source_ref:
          type: [string, "null"]