# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeadBody {
  name: string;
  phone?: string | null;
  email?: string | null;
  status?: string | null;
  answer_status?: string | null;
  capture_attempt_status?: string | null;
  lead_source?: string | null;
  ad_name?: string | null;
  referral_name?: string | null;
  activity_field?: string | null;
  followup_at?: string | null;
  followup_note?: string | null;
  rejection_reason?: string | null;
  rejection_reason_text?: string | null;
  first_deal_amount?: string | null;
  task_text?: string | null;
}