# lib/api-zod/src/generated/types/crmLead.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CrmLead {
  id: string;
  name: string;
  phone_e164: string;
  /** @nullable */
  phone_raw?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  sales_rep_id?: string | null;
  status_code: string;
  is_active_customer: boolean;
  /** @nullable */
  answer_status?: string | null;
  capture_attempts: number;
  /** @nullable */
  rejection_reason_code?: string | null;
  /** @nullable */
  rejection_detail?: string | null;
  pending_reassignment: boolean;
  /** @nullable */
  legacy_lead_id?: string | null;
  /** @nullable */
  linked_customer_id?: string | null;
  source: string;
  /** @nullable */
  source_ref?: string | null;
  created_at: Date;
  updated_at: Date;
  /** @nullable */
  deleted_at?: Date | null;
}